pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Edward"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "^2.0.1"
numpy = "^1.24.3"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"
torch = "^2.0.1"
skorch = "^0.12.1"
pykeops = "^2.1.2"
pyyaml = "^6.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
cmake = "^3.26.3"
scikit-survival = "^0.20.0"
matplotlib = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
